+++ Components/Comments.jsx
@@ -1,0 +1,7 @@
+import CommentsCard from "./CommentsCard"
+
+export default function Comments(){
+
+<CommentsCard comments={comments} />
+
+}
