+++ src/index.css
@@ -78,18 +78,22 @@ h1 {
   color: white; 
 }
 
+/*Articles*/
+
 .article-card {
   border: 1px solid #ddd;
   border-radius: 8px;
+  box-shadow: 2px 2px 10px rgba(0, 0, 0, 0.1);
   padding: 15px;
   margin-bottom: 20px;
   background: white;
-  box-shadow: 2px 2px 10px rgba(0, 0, 0, 0.1);
+  
  
     width: 800px;
     max-width: 100%;
-    word-wrap: break-word; 
-    overflow: hidden;
+    
+    
+    align-items: center;
   
 }
 
@@ -99,58 +103,50 @@ h1 {
   white-space: normal; 
   overflow-wrap: break-word;
 }
-.comments-section {
-  margin-top: 20px;
-  box-shadow: 2px 2px 10px rgba(0, 0, 0, 0.1);
+
+
+.article-body {
   border: 1px solid #ddd;
   border-radius: 8px;
+  padding: 40px;
+  font-size: 1rem;
+  color: #333;
+  justify-content: center;
+  align-items: center;
   
 }
 
-.comment {
-  margin-bottom: 20px;
-}
-
-.comment-details {
-  border-top: 1px solid #ddd;
- 
- align-items: center;
-  border-radius: 2px;
-  display: flex;
-  justify-content: space-between;
-  gap: 5px;
-}
-
-.comment-details p, .comment-details strong {
-  flex: 1;
-}
-
-.comment-details p {
-  text-align: center;
-}
-
-.article-body {
-  font-size: 1rem;
-  color: #333;
-}
-
 .meta-item {
+  opacity: 50%;
+  border: 1px solid #ddd;
+  border-radius: 8px;
+  box-shadow: 0px 3px 0px rgba(0, 0, 0, 0.05);
+  border: 1px;
+  display: flex;
+  justify-content: center; 
+  align-items: center;
+  min-width: 150px;
+}
+
+.meta-item article{
+  
+  justify-content: space-between;
+  border: 1px;
   display: flex;
   justify-content: center;
   align-items: center;
-  min-width: 150px;
+  opacity: 50%;
 }
 
 .article-meta {
   display: grid;
   grid-template-columns: repeat(4, 1fr);
-  gap: 10px;
+  gap: 5px;
   margin-top: 10px;
   text-align: center;
-  width: 400px;
+  
   max-width: 100%;
-  margin-left: auto;
-  margin-right: auto;
+  
 }
 
 .meta-column {
@@ -167,6 +163,50 @@ h1 {
   border-top: 1px solid #ddd;
   padding-top: 10px;
 }
+
+
+
+
+.comments-section {
+  margin-top: 20px;
+  box-shadow: 2px 2px 10px rgba(0, 0, 0, 0.1);
+  border: 1px solid #ddd;
+  
+  border-radius: 8px;
+  width: 830px;
+}
+
+.comment {
+  margin-bottom: 20px;
+}
+
+.comment-details {
+  border-top: 1px solid #ddd;
+ 
+ align-items: center;
+  border-radius: 2px;
+  display: flex;
+  justify-content: space-between;
+  gap: 2px;
+}
+
+.comment-details p, .comment-details strong {
+  flex: 1;
+  
+}
+
+.comment-details p {
+  text-align: center;
+  opacity: 50%;
+}
+
+.comment-body {
+
+  padding-left: 5%;
+  padding-right: 5%;
+  text-align: center;
+}
+
 
 .header-buttons {
   background: white;
